bug: 91793
CountObjects slow for Domain Admin.
If have more than 100 domains, use 3 threads to do the counting concurrently.

getAllDomains now caches the domains as long as the cache is big enough 
i.e. Localconfig setting ldap_cache_domain_maxsize is larger than the number of domains
(This is what is recommended at http://wiki.zimbra.com/wiki/OpenLDAP_Performance_Tuning_8.0 - at least if the number of domains is less than 30000).

This speeds up later CountObjects calls - which is useful for Admin Console which does 5 CountObjectsRequests at startup.

Use JAXB in the handler.  Fixed up the JAXB to fit with actual handler behavior - can specify multiple domains.
https://reviewboard.eng.zimbra.com/r/4457/
JUDASPRIEST ---> main

https://bugzilla.zimbra.com/show_bug.cgi?id=91793

Copied from Perforce
 Change: 540479

+++ src/db/mysql/create_database.sql
@@ -26,7 +26,7 @@ CREATE TABLE IF NOT EXISTS ${DATABASE_NAME}.mail_item (
    prev_folders  TEXT,                       -- e.g. "101:2;110:5", before mod_metadata 101, this item was in folder 2, before 110, it was in 5
    index_id      INTEGER UNSIGNED,
    imap_id       INTEGER UNSIGNED,
-   date          INTEGER UNSIGNED NOT NULL,  -- stored as a UNIX-style timestamp
+   date          BIGINT UNSIGNED NOT NULL,  -- stored as a UNIX-style timestamp
    size          BIGINT UNSIGNED NOT NULL,
    locator       VARCHAR(1024),
    blob_digest   VARCHAR(44) BINARY,         -- reference to blob
@@ -40,7 +40,7 @@ CREATE TABLE IF NOT EXISTS ${DATABASE_NAME}.mail_item (
    name          VARCHAR(255),               -- namespace entry for item (e.g. tag name, folder name, document filename)
    metadata      MEDIUMTEXT,
    mod_metadata  INTEGER UNSIGNED NOT NULL,  -- change number for last row modification
-   change_date   INTEGER UNSIGNED,           -- UNIX-style timestamp for last row modification
+   change_date   BIGINT UNSIGNED,           -- UNIX-style timestamp for last row modification
    mod_content   INTEGER UNSIGNED NOT NULL,  -- change number for last change to "content" (e.g. blob)
    uuid          VARCHAR(127),               -- e.g. "d94e42c4-1636-11d9-b904-4dd689d02402"
 
@@ -69,7 +69,7 @@ CREATE TABLE IF NOT EXISTS ${DATABASE_NAME}.mail_item_dumpster (
    prev_folders  TEXT,                       -- e.g. "101:2;110:5", before mod_metadata 101, this item was in folder 2, before 110, it was in 5
    index_id      INTEGER UNSIGNED,
    imap_id       INTEGER UNSIGNED,
-   date          INTEGER UNSIGNED NOT NULL,  -- stored as a UNIX-style timestamp
+   date          BIGINT UNSIGNED NOT NULL,  -- stored as a UNIX-style timestamp
    size          BIGINT UNSIGNED NOT NULL,
    locator       VARCHAR(1024),
    blob_digest   VARCHAR(44) BINARY,         -- reference to blob
@@ -83,7 +83,7 @@ CREATE TABLE IF NOT EXISTS ${DATABASE_NAME}.mail_item_dumpster (
    name          VARCHAR(255),               -- namespace entry for item (e.g. tag name, folder name, document filename)
    metadata      MEDIUMTEXT,
    mod_metadata  INTEGER UNSIGNED NOT NULL,  -- change number for last row modification
-   change_date   INTEGER UNSIGNED,           -- UNIX-style timestamp for last row modification
+   change_date   BIGINT UNSIGNED,           -- UNIX-style timestamp for last row modification
    mod_content   INTEGER UNSIGNED NOT NULL,  -- change number for last change to "content" (e.g. blob)
    uuid          VARCHAR(127),               -- e.g. "d94e42c4-1636-11d9-b904-4dd689d02402"
 
@@ -105,14 +105,14 @@ CREATE TABLE IF NOT EXISTS ${DATABASE_NAME}.revision (
    mailbox_id    INTEGER UNSIGNED NOT NULL,
    item_id       INTEGER UNSIGNED NOT NULL,
    version       INTEGER UNSIGNED NOT NULL,
-   date          INTEGER UNSIGNED NOT NULL,  -- stored as a UNIX-style timestamp
+   date          BIGINT UNSIGNED NOT NULL,  -- stored as a UNIX-style timestamp
    size          BIGINT UNSIGNED NOT NULL,
    locator       VARCHAR(1024),
    blob_digest   VARCHAR(44) BINARY,         -- reference to blob
    name          VARCHAR(255),               -- namespace entry for item (e.g. tag name, folder name, document filename)
    metadata      MEDIUMTEXT,
    mod_metadata  INTEGER UNSIGNED NOT NULL,  -- change number for last row modification
-   change_date   INTEGER UNSIGNED,           -- UNIX-style timestamp for last row modification
+   change_date   BIGINT UNSIGNED,           -- UNIX-style timestamp for last row modification
    mod_content   INTEGER UNSIGNED NOT NULL,  -- change number for last change to "content" (e.g. blob)
 
    PRIMARY KEY (mailbox_id, item_id, version),
@@ -125,14 +125,14 @@ CREATE TABLE IF NOT EXISTS ${DATABASE_NAME}.revision_dumpster (
    mailbox_id    INTEGER UNSIGNED NOT NULL,
    item_id       INTEGER UNSIGNED NOT NULL,
    version       INTEGER UNSIGNED NOT NULL,
-   date          INTEGER UNSIGNED NOT NULL,  -- stored as a UNIX-style timestamp
+   date          BIGINT UNSIGNED NOT NULL,  -- stored as a UNIX-style timestamp
    size          BIGINT UNSIGNED NOT NULL,
    locator       VARCHAR(1024),
    blob_digest   VARCHAR(44) BINARY,         -- reference to blob
    name          VARCHAR(255),               -- namespace entry for item (e.g. tag name, folder name, document filename)
    metadata      MEDIUMTEXT,
    mod_metadata  INTEGER UNSIGNED NOT NULL,  -- change number for last row modification
-   change_date   INTEGER UNSIGNED,           -- UNIX-style timestamp for last row modification
+   change_date   BIGINT UNSIGNED,           -- UNIX-style timestamp for last row modification
    mod_content   INTEGER UNSIGNED NOT NULL,  -- change number for last change to "content" (e.g. blob)
 
    PRIMARY KEY (mailbox_id, item_id, version),
@@ -209,7 +209,7 @@ CREATE UNIQUE INDEX i_item_id ON ${DATABASE_NAME}.appointment_dumpster (mailbox_
 CREATE TABLE IF NOT EXISTS ${DATABASE_NAME}.tombstone (
    mailbox_id  INTEGER UNSIGNED NOT NULL,
    sequence    INTEGER UNSIGNED NOT NULL,  -- change number for deletion
-   date        INTEGER UNSIGNED NOT NULL,  -- deletion date as a UNIX-style timestamp
+   date        BIGINT UNSIGNED NOT NULL,  -- deletion date as a UNIX-style timestamp
    type        TINYINT,                    -- 1 = folder, 3 = tag, etc.
    ids         TEXT,
